[200~About page created with text explaining the project and a link back to the main page.
Contact page created with a link to the creator's Github and a link
back to the main page.

Main page created in the index.html file where links to all the recipes,
and misceallaneous pages.

Three recipes have been added with different descriptions and steps
for preparations.

So far I have practiced using tags for titles, paragraphs, links,
ordered lists,and unordered lists.~